remove get-storage-verified
