+++ core-primitives/ocall-api/Cargo.toml
@@ -7,22 +7,27 @@ resolver = "2"
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
+derive_more = { version = "0.99.5" }
+
+# sgx deps
+sgx_types = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 
 # substrate deps
+sp-core = { version = "6.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.19"}
 sp-runtime = { version = "6.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.19"}
 sp-std = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.19"}
 
-# sgx-deps
-sgx_types = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
-
 # local deps
+itp-storage = { path = "../storage", default-features = false }
 itp-types = { path = "../types", default-features = false }
 
 [features]
 default = ["std"]
 std = [
     "codec/std",
+    "sp-core/std",
     "sp-runtime/std",
     "sp-std/std",
+    "itp-storage/std",
     "itp-types/std",
 ]
