+++ core-primitives/test/src/mock/onchain_mock.rs
@@ -20,7 +20,6 @@ use codec::{Decode, Encode};
 use core::fmt::Debug;
 use itp_ocall_api::{EnclaveAttestationOCallApi, EnclaveMetricsOCallApi, EnclaveSidechainOCallApi};
 use itp_storage::StorageEntryVerified;
-use itp_storage_verifier::{GetStorageVerified, Result};
 use itp_teerex_storage::{TeeRexStorage, TeerexStorageKeys};
 use itp_types::{BlockHash, Enclave, ShardIdentifier};
 use sgx_types::{
@@ -64,36 +63,6 @@ impl OnchainMock {
 
 	pub fn get(&self, key: &[u8]) -> Option<&Vec<u8>> {
 		self.inner.get(key)
-	}
-}
-
-impl GetStorageVerified for OnchainMock {
-	fn get_storage_verified<H: HeaderT<Hash = H256>, V: Decode>(
-		&self,
-		storage_hash: Vec<u8>,
-		_header: &H,
-	) -> Result<StorageEntryVerified<V>> {
-		let value = self
-			.get(&storage_hash)
-			.map(|val| Decode::decode(&mut val.as_slice()))
-			.transpose()?;
-
-		Ok(StorageEntryVerified::new(storage_hash, value))
-	}
-
-	fn get_multiple_storages_verified<H: HeaderT<Hash = H256>, V: Decode>(
-		&self,
-		storage_hashes: Vec<Vec<u8>>,
-		_header: &H,
-	) -> Result<Vec<StorageEntryVerified<V>>> {
-		let mut entries = Vec::with_capacity(storage_hashes.len());
-		for hash in storage_hashes.into_iter() {
-			let value =
-				self.get(&hash).map(|val| Decode::decode(&mut val.as_slice())).transpose()?;
-
-			entries.push(StorageEntryVerified::new(hash, value))
-		}
-		Ok(entries)
 	}
 }
 
@@ -160,22 +129,46 @@ impl EnclaveMetricsOCallApi for OnchainMock {
 	}
 }
 
-// We cannot implement EnclaveOnChainOCallApi specifically here, because OnchainMock already
-// implements `GetStorageVerified`. And all implementers of `EnclaveOnChainOCallApi` automatically
-// implement GetStorageVerified too (-> see `core-primitives/storage-verified/src/lib.rs`),
-// so it results in duplicate implementations.
-// impl EnclaveOnChainOCallApi for OnchainMock {
-// 	fn send_to_parentchain(&self, _extrinsics: Vec<OpaqueExtrinsic>) -> SgxResult<()> {
-// 		Ok(())
-// 	}
-//
-// 	fn worker_request<V: Encode + Decode>(
-// 		&self,
-// 		_req: Vec<WorkerRequest>,
-// 	) -> SgxResult<Vec<WorkerResponse<V>>> {
-// 		Ok(Vec::new())
-// 	}
-// }
+impl EnclaveOnChainOCallApi for OnchainMock {
+	fn send_to_parentchain(&self, _extrinsics: Vec<OpaqueExtrinsic>) -> SgxResult<()> {
+		Ok(())
+	}
+
+	fn worker_request<V: Encode + Decode>(
+		&self,
+		_req: Vec<WorkerRequest>,
+	) -> SgxResult<Vec<WorkerResponse<V>>> {
+		Ok(Vec::new())
+	}
+
+	fn get_storage_verified<H: HeaderT<Hash = H256>, V: Decode>(
+		&self,
+		storage_hash: Vec<u8>,
+		_header: &H,
+	) -> Result<StorageEntryVerified<V>> {
+		let value = self
+			.get(&storage_hash)
+			.map(|val| Decode::decode(&mut val.as_slice()))
+			.transpose()?;
+
+		Ok(StorageEntryVerified::new(storage_hash, value))
+	}
+
+	fn get_multiple_storages_verified<H: HeaderT<Hash = H256>, V: Decode>(
+		&self,
+		storage_hashes: Vec<Vec<u8>>,
+		_header: &H,
+	) -> Result<Vec<StorageEntryVerified<V>>> {
+		let mut entries = Vec::with_capacity(storage_hashes.len());
+		for hash in storage_hashes.into_iter() {
+			let value =
+				self.get(&hash).map(|val| Decode::decode(&mut val.as_slice())).transpose()?;
+
+			entries.push(StorageEntryVerified::new(hash, value))
+		}
+		Ok(entries)
+	}
+}
 
 pub fn validateer_set() -> Vec<Enclave> {
 	let default_enclave = Enclave::new(
