+++ sidechain/consensus/aura/src/lib.rs
@@ -31,7 +31,6 @@ extern crate sgx_tstd as std;
 
 use core::marker::PhantomData;
 use itc_parentchain_block_import_dispatcher::triggered_dispatcher::TriggerParentchainBlockImport;
-use itp_storage_verifier::GetStorageVerified;
 use itp_time_utils::duration_now;
 use its_consensus_common::{Environment, Error as ConsensusError, Proposer};
 use its_consensus_slots::{SimpleSlotWorker, Slot, SlotInfo};
