+++ sidechain/validateer-fetch/src/validateer.rs
@@ -17,7 +17,6 @@
 
 use crate::error::{Error, Result};
 use frame_support::ensure;
-use itp_storage_verifier::GetStorageVerified;
 use itp_teerex_storage::{TeeRexStorage, TeerexStorageKeys};
 use itp_types::Enclave;
 use sp_core::H256;
